+++ RELEASE_NOTES.md
@@ -1,14 +1,14 @@
 Release Notes
 =============
-
 ## 1.7.1
 * Event Hub: Don't create the `$Default` consumer group explicitly. It will automatically be created by Azure when the resource is created.
+* Static Web Apps: App Setting support.
 
 ## 1.7.0
 * Azure CLI: Escape parameters passed to the az deployment command (breaking change). Any previously escaped parameters need to be unescaped before passing to the tryValidate, tryWhatIf, tryExecute, whatIf and execute functions.
 
 ## 1.6.35
-* Container Apps: Support for Managed Identities 
+* Container Apps: Support for Managed Identities
 * Logic Apps: Basic support for logic apps. These will require the logic app code to be supplied either directly or via file path.
 
 ## 1.6.34