+++ RELEASE_NOTES.md
@@ -1,6 +1,5 @@
 Release Notes
 =============
-
 ## 1.7.1
 * App Insights: Add ConnectionString member.
 * Communication Services: **Breaking Changes**: Clean up and fix issues regarding naming and Location.
@@ -8,12 +7,13 @@ Release Notes
 * Event Hub: Don't create the `$Default` consumer group explicitly. It will automatically be created by Azure when the resource is created.
 * SignalR: Add ConnectionString member.
 * SignalR: **Breaking Change**: Bug fix - Key now returns Key, not ConnectionString.
+* Static Web Apps: App Setting support.
 
 ## 1.7.0
 * Azure CLI: Escape parameters passed to the az deployment command (breaking change). Any previously escaped parameters need to be unescaped before passing to the tryValidate, tryWhatIf, tryExecute, whatIf and execute functions.
 
 ## 1.6.35
-* Container Apps: Support for Managed Identities 
+* Container Apps: Support for Managed Identities
 * Logic Apps: Basic support for logic apps. These will require the logic app code to be supplied either directly or via file path.
 
 ## 1.6.34